Permit paths to start and end on aliases of REGs
This patch tries to resolve a problem that occurs when modules with
registered ports are inlined by Verilator, causing the variable
reference that is the target of a non-blocking assignment, to not
correspond to the original source. Correct dependencies are maintained
with alias assignments to variables of the original name.

This patch tries to improve the mapping of paths to the original source
by:

- Introducing alias vars on destination registers. Previously, they were
  all being moved to the source registers. If we have

    REG -> ASSIGN_ALIAS -> VAR

  Then duplicate ASSIGN_ALIAS and VAR so there is:

    SRC_REG <- ASSIGN_ALIAS <- VAR
    DST_REG -> ASSIGN_ALIAS -> VAR

- Allowing paths to start and finish on aliases of source and destination
  registers respectively.