Add option to allow paths to traverse through registers
This is to enable connectivity checks to be written using the library.

The feature is implemeneted by retaining the edges from (destination)
registers to the nodes connected by out edges, but marking them as
'through register'. Graph DFS traversals are then performed on a
filtered graph, with an edge predicate selecting the through-register
edges or not.