Prevent multi-edges from occurring
This was causing a problem with the marking of edges that are 'through'
a register. Using boost::edge() to access an edge was always returning
the first edge when there were multiple, leaving the other edge(s) not
marked as 'through register'. Previously, removing the edge with
boost::remove_edge() removed all edges between two vertices so had the
desired effect.